+++ src/lib_injector/injector_server.ml
@@ -84,6 +84,8 @@ module Parameters :
 
   let safety_guard _ = None
 
+  let max_batch_length _state = None
+
   let persist_operation _ = true
 
   let retry_unsuccessful_operation _node_ctxt (_op : Operation.t) ?reason:_